doc_assistant/requirements.txt
requests
pymupdf
python-docx
pathlib
